Split formatter into module with logical subfiles
- formatter.rs: struct, public API, fmt_node dispatch, CST query helpers
- formatter/items.rs: imports, const, function, type declarations
- formatter/expr.rs: all expression formatting (binary, pipe, match, etc.)
- formatter/jsx.rs: JSX element/prop/children formatting
- formatter/tests.rs: all 17 formatter tests
